build: gate toolruntime tracing behind a build tag
